speed up spanplus
Former-commit-id: e9638a308037df1d1036b32587339e7f13ab8401
Former-commit-id: 769bdd0330a52b997ba37d12c9fcc46f845824a3

+++ backend/src/pytorch/spandrel/architectures/__arch_helpers/dysample.py
@@ -89,6 +89,6 @@ class DySample(nn.Module):
         )
 
         if self.end_convolution:
-            output = self.end_conv(output)
+            output = self.end_conv(output.to(x.dtype)).to(x.dtype)
 
-        return output
+        return output.to(x.dtype)
